+++ math/common/CMakeLists.txt
@@ -12,7 +12,6 @@ algebra_add_library(algebra_common_math common_math
    "include/algebra/coordinates/cylindrical2.hpp"
    "include/algebra/coordinates/line2.hpp"
    "include/algebra/coordinates/polar2.hpp"
-   "include/algebra/math/common.hpp"
-   "include/algebra/math/operators.hpp")
+   "include/algebra/math/common.hpp")
 target_link_libraries(algebra_common_math
    INTERFACE algebra::common)
